Refactor CMake configuration and move CLI module
