+++ src/CMakeLists.txt
@@ -1,9 +1,10 @@
 cmake_minimum_required(VERSION 3.20)
 project(RSA-Encryptor CXX)
 
-add_executable(RSA-Encryptor main.cpp cli.cpp)
-
 add_subdirectory(core)
 add_subdirectory(vec)
+add_subdirectory(cli)
 
-target_link_libraries(RSA-Encryptor PRIVATE RSA-Vec)
+add_executable(RSA-Encryptor main.cpp)
+
+target_link_libraries(RSA-Encryptor PRIVATE RSA-Vec RSA-Core RSA-CLI)
