+++ src/cli/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(RSA-CLI
+    cli.cpp
+)
